refactor: move IOU types from index.ts into their function files
Completes the IOU/index.ts cleanup series by relocating the 4 types
defined inline and dropping the 3 type re-exports:

Types inlined into their function files:
- ReplaceReceipt → Receipt.ts
- StartSplitBilActionParams → Split.ts
- UpdateMoneyRequestData<TKey> → UpdateMoneyRequest.ts

IOURequestType is the canonical CONST alias, so callers now import it
directly from @src/CONST (which already exports it).

Re-exports dropped from index.ts — callers updated to import directly
from the actual type files:
- GpsPoint → ./types/TrackExpenseTransactionParams (GPSPoint)
- RequestMoneyParticipantParams → ./types/RequestMoneyParticipantParams
- BaseTransactionParams → ./types/BaseTransactionParams

After this PR, IOU/index.ts contains only Onyx vars and getter functions —
no types, no setters. File shrinks to 235 lines.
